Added more 3rd requiremnet stuff

+++ ENSE496-A4/ENSE496-A4/Program.cs
@@ -68,7 +68,7 @@ namespace ENSE496_A4
 
         public ulong PowerFunct(ulong a, ulong b, ulong P)
         {
-            BigInteger temp = 0;
+            ulong temp = 0;
             ulong i;
             if (a == b)
                 return a;
